% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_permissions_RLDNe.R
\name{fix_permissions_RLDNe}
\alias{fix_permissions_RLDNe}
\title{A function that will generate or execute a system command that changes file permissions of NeEstimator V2.1 executables on Linux or Mac OSs.}
\usage{
fix_permissions_RLDNe(execute = F)
}
\arguments{
\item{execute}{Logical (Default: FALSE), if true a system command will be executed to add execuatable permissions to distributed executable files. If False, the necessary command is provided to the user.}
}
\description{
A function that will generate or execute a system command that changes file permissions of NeEstimator V2.1 executables on Linux or Mac OSs.
}
\author{
Zak Robinson, Contact: zachary.robinson(at)umontana.com
}
