% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypes2alleles.R
\name{genotypes2alleles}
\alias{genotypes2alleles}
\title{Convert wide format Genotypes to wider format Alleles}
\usage{
genotypes2alleles(df, loci, fixed_A = NULL, delim = NULL, name_sep = ".")
}
\arguments{
\item{df}{Data frame that contains genotypic data and metadata}

\item{loci}{Numeric column positions or names of loci}

\item{fixed_A}{The number of characters of alleles in input. A diploid genotype "AT" has a fixed_A=1. Only use when the delimiter between alleles is "".}

\item{delim}{A non-zero length character string separating alleles in input. For example, ":" for "A:T" or "-" 114-116"}

\item{name_sep}{The desired locus allele name separator for output column names e.g., "_" for Loc1_A1. Defaults to "."}
}
\value{
Data frame in allele per column format
}
\description{
A function that will return a data frame in allele per column format from genotype per column format.
}
\details{
This function is meant to conveniently wrap tidyr and dplyr operations to convert data between two common wide formats. At present, it's just set up for diploid data.
}
\examples{
data("wgp_example")
alleles <- genotypes2alleles(wgp_example,loci = grep("locus",colnames(wgp_example)),fixed_A = 1)
}
\seealso{
\code{\link{alleles2genotypes}}
}
\author{
Zak Robinson, Contact: <zrobinson@critfc.org>
}
