% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeV2_LDNE_create_zlr.R
\name{NeV2_LDNe_create}
\alias{NeV2_LDNe_create}
\title{Write LDNe Parameter File}
\usage{
NeV2_LDNe_create(input_file, param_file, NE_out_file, matingsystem = 1,
  crit_vals = c(0.02, 0.05, 0.1))
}
\arguments{
\item{input_file}{Genepop format file containing genotypic data}

\item{param_file}{Desired name of parameter file produced}

\item{matingsystem}{0: Random mating, 1: Monogamy (LD method)}

\item{crit_vals}{minimum allele frequency cutoff. Defaults to c(0.02,0.05,0.1).}

\item{Ne_out_file}{Desired name of output from LDNe}
}
\value{
Path of created LDNe param file
}
\description{
This function writes a parameter for LD Method in NeEstmator V2.1
}
\author{
Zak Robinson, Contact: zachary.robinson(at)umontana.com
}
