% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportGenepop_RLDNe.R
\name{print.RLDNe_data}
\alias{print.RLDNe_data}
\title{print method for RLDNe_data}
\usage{
\method{print}{RLDNe_data}(x, ...)
}
\arguments{
\item{x}{an RLDNe_data object}

\item{...}{ignored}
}
\description{
print method for RLDNe_data
}
