% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_LDNeOutFile.R
\name{read_LDNeOutFile}
\alias{read_LDNeOutFile}
\title{Read LDNe Tabular Output}
\usage{
read_LDNeOutFile(x, PopDecode = T)
}
\arguments{
\item{x}{An RLDNe object or file path to tabular LDNe output.}

\item{PopDecode}{Attempt to recode populations to original name. Default=T}
}
\value{
data.frame of parsed LDNe output
}
\description{
This function parses the tabular output from LDNe
}
