% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLDNe_tab.R
\name{readLDNe_tab}
\alias{readLDNe_tab}
\title{Write LDNe Parameter File}
\usage{
readLDNe_tab(path)
}
\arguments{
\item{path}{path to tabular format LDNe output}
}
\value{
Path of submitted LDNe output and dataframe of parsed data
}
\description{
This function parses the tabular output from LDNe
}
\author{
Zak Robinson, Contact: zachary.robinson(at)umontana.com
}
