% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_LDNe.R
\name{run_LDNe}
\alias{run_LDNe}
\title{Function to Run LDNe from NeEstimator 2.1}
\usage{
run_LDNe(LDNe_params)
}
\arguments{
\item{LDNe_params}{LDNe parameter file, best to use output from NEV2_LDNE_create_zlr}
}
\description{
Function to Run LDNe from NeEstimator 2.1
}
\author{
Zak Robinson, Contact: zachary.robinson(at)umontana.com
}
