% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_LDNe.R
\name{run_LDNe_32bit}
\alias{run_LDNe_32bit}
\title{Function to Run LDNe from NeEstimator 2.1 using 32-bit executables}
\usage{
run_LDNe_32bit(x)
}
\arguments{
\item{x}{An RLDNe object or file path to parameter file}
}
\description{
Function to Run LDNe from NeEstimator 2.1 using 32-bit executables
}
\author{
Zak Robinson, Contact: <zrobinson@critfc.org>
}
