% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleles2genotypes.R
\name{alleles2genotypes}
\alias{alleles2genotypes}
\title{Convert Alleles to Genotypes (hierfstat format)}
\usage{
alleles2genotypes(df, allele_cols, allelesAsIntegers = T)
}
\arguments{
\item{df}{Data frame that contains genotypic data and metadata}

\item{allele_cols}{Numeric column identifiers of loci}

\item{allelesAsIntegers}{For microsats with leading zeros use FALSE. If FALSE this will eliminate the loss of leading zeros, but will not be usable in 'hierfstat'. When TRUE the genotypes are converted to integers for 'hierfstat' package.}
}
\value{
Data frame in genotype per column format
}
\description{
A function that will return a data frame in genotype per column format from a data frame in allele per column format.
}
\examples{
data("wgp_example")
alleles2genotypes(wgp_example_2col,allele_cols = 3:ncol(wgp_example_2col),allelesAsIntegers = T)
}
\author{
Zak Robinson, Contact: zachary.robinson(at)umontana.com
}
